Extracted backend to separate project
