+++ Reserve/Reserve.Core/Features/Queue/QueueTicket.cs
@@ -1,6 +1,5 @@
 ﻿using System.ComponentModel.DataAnnotations;
-
-namespace Reserve.Models.Queue;
+namespace Reserve.Core.Features.Queue;
 
 public class QueueTicket
 {
