+++ Reserve/Reserve.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 17.8.34004.107
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Reserve", "Reserve\Reserve.csproj", "{E937F7A2-F9F6-4652-842E-834AC6A338F4}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Reserve.Core", "Reserve.Core\Reserve.Core.csproj", "{A6F9A5A6-7E64-464B-97E8-9DB505B5A684}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{E937F7A2-F9F6-4652-842E-834AC6A338F4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E937F7A2-F9F6-4652-842E-834AC6A338F4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E937F7A2-F9F6-4652-842E-834AC6A338F4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A6F9A5A6-7E64-464B-97E8-9DB505B5A684}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A6F9A5A6-7E64-464B-97E8-9DB505B5A684}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A6F9A5A6-7E64-464B-97E8-9DB505B5A684}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A6F9A5A6-7E64-464B-97E8-9DB505B5A684}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
