+++ Reserve/Reserve/Pages/Queue/QueueTicket.cshtml.cs
@@ -1,7 +1,6 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
-using Reserve.Models.Queue;
-using Reserve.Repositories;
+using Reserve.Core.Features.Queue;
 
 namespace Reserve.Pages.Queue;
